requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.1.0
attrs==19.3.0
black==19.10b0
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.4.2
Django==2.1.5
django-multiselectfield==0.1.8
enum34==1.1.6
flake8==3.6.0
future==0.17.1
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pathspec==0.7.0
pep8==1.7.1
Pillow==6.0.0
psycopg2==2.7.6.1
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
pylint==2.2.2
python-memcached==1.59
python-slugify==3.0.2
python-telegram-bot==11.1.0
pytz==2018.9
regex==2020.2.20
requests==2.21.0
rope==0.11.0
six==1.12.0
text-unidecode==1.2
toml==0.10.0
typed-ast==1.4.1
ujson==1.35
urllib3==1.24.1
vk-api==11.4.0
wincertstore==0.2
wrapt==1.10.11
